feat: add combo to notes; different algorithm for melody notes

+++ app/domains/maps/jobs/queues/queue.py
@@ -64,7 +64,7 @@ class ExtractMelodyJobPayload(BaseModel):
 
     job_id: str
     lane_count: int
-    bpm: float
+    bpm: int
 
 
 class KickDetectionResult(BaseModel):
